create add and remove functionalities
